+++ src/app/components/edition/edit-combat/edit-combat.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { IonicModule } from '@ionic/angular';
+import { EditCombatComponent } from './edit-combat.component';
+
+describe('EditPowersComponent', () => {
+  let component: EditCombatComponent;
+  let fixture: ComponentFixture<EditCombatComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [ EditCombatComponent ],
+      imports: [IonicModule.forRoot()]
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(EditCombatComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});