+++ src/app/services/json.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { GiftData } from '../model/giftData';
+import { PowersData } from '../model/powerData';
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +11,7 @@ export class JsonService {
 
   constructor(private http: HttpClient) { }
 
-  public getJsonData(jsonFile: string) {
+  public getGiftsDataFromJson(jsonFile: string) {
     return this.http.get(jsonFile).pipe(
       map(data => {
         return data as GiftData;
@@ -18,6 +19,11 @@ export class JsonService {
     );
   }
 
-
+  public getPowersDataFromJson(jsonFile: string) {
+    return this.http.get(jsonFile).pipe(
+      map(data => {
+        return data as PowersData;
+      })
+    );
+  }
 }
-