+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Aureo_App
 ## V1.0.5
 - Edit History directly in detail.
+- Combat equip catalog added.
+- Combat equip types added.
+- Damage calculated from weapon types.
+- Combat equipment added to wiki page.
+- Combat equipment detail alert added.
+- Combat equipment edition.
+- Some Normalization.
+- Json file resuming all powers.
+- Powers added to wiki page.
+- Powers detail alert added.
+- String no concat.
+- Powers edition.
 
 ## V1.0.4
 - Edit Afinities.